Support disabling compile-time HTTP requests
This adds a new configuration flag, `:fetch_latest`, which entirely disables the HTTP request at compile-time that tries to get the latest `public_suffix_list.dat`. This behavior was (sort of) achievable without this flag, by configuring the `:public_suffix_list_url` to the empty charlist, but doing so produced a warning on each compile. With the new flag, there's no error.

This is desirable for our use-case for a couple reasons:

1. We want the development build to match as closely as possible what we'll deploy in prod
2. We don't want the possibility of failing production builds based on the availability of a third-party web host