Fix International Domain Names
I noticed that parsing didn't work with basically any Unicode domain.
After investigating I noticed that when fetching the latest list, erlang
was transforming the result to a char list and the unicode was lost in
the process.

This fix tells erlang to keep the response body as a binary, which
elixir already interprets correctly as UTF-8.